app/src/test/java/com/cshawn/codes/leetcodes/everyday/Q825Test.java
package com.cshawn.codes.leetcodes.everyday;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;
import java.util.stream.Stream;
import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author C.Shawn
 * @date 2021/12/27 9:45 上午
 */
class Q825Test {
    static class DataArgumentsProvider implements ArgumentsProvider {
        @Override
        public Stream<? extends Arguments> provideArguments(ExtensionContext context) throws Exception {
            return Stream.of(
                    Arguments.of(new int[]{16,16}, 2),
                    Arguments.of(new int[]{16,17,18}, 2),
                    Arguments.of(new int[]{20,30,100,110,120}, 3)
            );
        }
    }

    @ParameterizedTest
    @ArgumentsSource(DataArgumentsProvider.class)
    void numFriendRequests(int[] ages, int result) {
        assertEquals(result, new Q825().numFriendRequests1(ages));
        assertEquals(result, new Q825().numFriendRequests(ages));
    }
}